Tina Reis's avatar include og:locale in open graph meta data
The open graph protocol enables web pages to show up as rich objects in
a social graph, see: https://ogp.me/
Which basically means that when someone shares a link of a webpage on
Facebook, it will look like a page on Facebook instead of an external
web page.
Hugo has an internal template for Open Graph meta data:
https://gohugo.io/templates/internal/#open-graph
It is used in Terrassa theme.
However, I wanted to include the optional meta tag og:locale, to signal
to social media platforms which language the page is in.
Hugo's template includes og:locale, but it is set using page parameters,
 not site parameters, see code here on line 30:
 https://github.com/gohugoio/hugo/blob/master/tpl/tplimpl/embedded/templates/opengraph.html
This doesn't make sense to me, because then we'd have to define the
locale parameter in the front matter of every single page for it to render. We could do that automatically by adding a line in `archetypes` default, which defines the default parameters that are added to each page, but still this means an additional line in every page that we create.
So I decided to manually add og:locale meta tag in head.html.
I can come back to this later, if I ever figure out why Hugo does it
that way, and change it but this works for now.