Fixed link to home page in header on translated pages
There is a link to the home page in the header, which is generated by
the .Site.BaseURL variable in header.html layout. This link refers to
the home page of the defaultContentLanguage everywhere, also on
translated pages.

Expected behavior: link to to the translated home page

There can only be one BaseURL, since it is a global variable that can't
be overwritten by language parameters.

According to this:
https://jibecfed.fedorapeople.org/blog-hugo/en/2018/11/building-a-multilingual-static-website-with-hugo/
a fix should be using {{.Site.BaseURL | relLangURL }}

but it is not working. I replaced the .Site.BaseURL with "/" (which is
what is currently the baseURL), and it works.

the relLangURL function seems to be buggy in Hugo: https://github.com/gohugoio/hugo/pull/5369

But I haven't found a solution yet.